% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createShiny.R
\name{createShiny}
\alias{createShiny}
\title{createShiny}
\usage{
createShiny(name = NA)
}
\arguments{
\item{name}{(string) the name of the project.

         If you do not provide a value, then it will use the current working
         directory as the project's root directory.}
}
\description{
Create shiny project.
}
\details{
If you specify a value for the "name",  then it will create a
         subdirectory of the same name which acts as the projects root
         directory. Otherwise, it will use the current working directory
         as the root directory of the project.
}
\note{
This function has only been tested on Linux, it might not work on other
      operating systems yet.
}
\examples{
# Convert the current directory to a Shiny Project App
createShiny()

# Create a Shiny Project App called "myApp" in a subdirectory called "myApp"
createShiny("myApp")
}
\author{
Ronny Restrepo
}
\keyword{app}
\keyword{app,}
\keyword{shiny}
\keyword{shiny,}

